# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Lifio API
  description: |
    Backend API for the Lifio personal time-tracking iOS app.

    ## Auth flow (passwordless email-OTP)
    1. **Request OTP** — `POST /api/v1/auth/otp/request` with an email address. Always returns 202
       (prevents user enumeration). The server upserts the user, generates a 6-digit code,
       and sends it via email (console in dev/test, AWS SES in production).
    2. **Verify OTP** — `POST /api/v1/auth/otp/verify` with the email and 6-digit code. On success,
       returns a JWT access token (15 min TTL) and an opaque refresh token.
    3. **Refresh** — `POST /api/v1/auth/refresh` with the refresh token and `X-Device-Id`.
       Issues a new token pair within the sending device's refresh family and revokes the
       old token. Reuse of a revoked token revokes only that device's family — other
       devices' sessions keep working.
    4. **Logout** — `POST /api/v1/auth/logout` with a Bearer access token and `X-Device-Id`.
       Revokes only the calling device's refresh family; other devices' sessions remain valid.
    5. **Me** — `GET /api/v1/auth/me` with a Bearer access token. Returns the user profile.

    ## Auth flow (Sign in with Apple)
    - **Apple** — `POST /api/v1/auth/apple` with Apple's identity-token JWT (RS256). The server
      verifies it against Apple's JWKS (iss/aud/exp) and upserts a user keyed by Apple's
...
       The route is always registered; without `APPLE_CLIENT_ID` it returns 503
...

...

...

...

...
servers:
...
paths:
...
  /api/v1/auth/otp/request:
    post:
      operationId: requestOTP
      summary: Request an OTP code
      description: |
        Sends a 6-digit OTP code to the given email address. The user account is
        created automatically if it does not exist. Always returns HTTP 202 to
        prevent user enumeration — a 202 response does NOT guarantee that an email
        was sent (internal errors are silently swallowed).
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
            example:
              email: user@example.com
      responses:
        "202":
          description: Request accepted (email may or may not have been sent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: invalid_body
                  message: Invalid request body
        "429":
          description: Rate limited (3 req/min per IP+email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.
  /api/v1/auth/otp/verify:
    post:
      operationId: verifyOTP
      summary: Verify an OTP code
      description: |
        Verifies the 6-digit OTP code for the given email. On success, marks the
        user as email-verified and returns a JWT access token + opaque refresh token.
        The refresh token is recorded against the `X-Device-Id` device's refresh family.
      tags:
        - Auth
      parameters:
        - $ref: "#/components/parameters/DeviceIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
            example:
              email: user@example.com
              code: "123456"
      responses:
        "200":
          description: OTP verified successfully — tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request body (invalid email or code format) or missing device header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_body:
                  summary: Invalid email or code format
                  value:
                    error:
                      code: invalid_body
                      message: Invalid email or code
                invalid_request:
                  summary: Missing X-Device-Id header
                  value:
                    error:
                      code: invalid_request
                      message: X-Device-Id header is required
        "401":
          description: Invalid, expired, or exhausted OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_otp:
                  summary: Invalid code
                  value:
                    error:
                      code: invalid_otp
                      message: Invalid or expired code
                otp_expired:
                  summary: Code expired
                  value:
                    error:
                      code: otp_expired
                      message: Code has expired. Request a new one.
                otp_attempts_exceeded:
                  summary: Too many failed attempts
                  value:
                    error:
                      code: otp_attempts_exceeded
                      message: Too many incorrect attempts. Request a new code.
        "429":
          description: Rate limited (5 req/min per IP+email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many attempts. Please try again later.
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      description: |
        Exchanges a refresh token for a new access + refresh token pair.
        The old refresh token is revoked and the new one stays within the
        `X-Device-Id` device's refresh family. Rotation never affects other
        devices' sessions. There is deliberately NO device limit: no 409
        `device_limit` response exists (device quota counting is deferred to
        issue #46 — do not file its absence as a bug).
      tags:
        - Auth
      parameters:
        - $ref: "#/components/parameters/DeviceIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refresh_token: abc123...
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request body or missing device header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_body:
                  summary: Refresh token missing or malformed
                  value:
                    error:
                      code: invalid_body
                      message: Refresh token is required
                invalid_request:
                  summary: Missing X-Device-Id header
                  value:
                    error:
                      code: invalid_request
                      message: X-Device-Id header is required
        "401":
          description: |
            Invalid, revoked (reuse), or TTL-expired refresh token. Reuse detection is
            scoped per-device: only the offending device's refresh family is revoked —
            other devices' sessions keep working. After a revoked or TTL-expired family,
            the device must re-authenticate (the client locks to the auth gate with all
            local data intact; nothing is wiped).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_refresh:
                  summary: Token not found or presented from the wrong device
                  value:
                    error:
                      code: invalid_refresh
                      message: Invalid refresh token
                token_reuse:
                  summary: Token reuse detected (offending device's family revoked)
                  value:
                    error:
                      code: token_reuse
                      message: Token has been revoked. The device's sessions have been invalidated.
                refresh_expired:
                  summary: Refresh family TTL exhausted (re-authenticate, local data intact)
                  value:
                    error:
                      code: refresh_expired
                      message: Refresh token has expired. Please sign in again.
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    DeviceIdHeader:
      name: X-Device-Id
      in: header
      required: true
      description: |
        Stable client-generated device identifier (UUID). Persisted in the iOS Keychain
        across app reinstalls; minted fresh only when the device fingerprint changes
        (restore-on-new-phone). Scopes refresh-token families, reuse detection, and
        logout to a single device: revocation never touches other devices' sessions.
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/api/v1/auth/otp/verify`, `/api/v1/auth/apple`, or `/api/v1/auth/refresh`.
        Format: `Authorization: Bearer <token>`
...
  schemas:
...
    OtpRequest:
      type: object
      description: Request body for requesting an OTP code.
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: User's email address
          example: user@example.com
    OtpVerify:
      type: object
      description: Request body for verifying an OTP code.
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: User's email address
          example: user@example.com
        code:
          type: string
          minLength: 6
          maxLength: 6
          pattern: "^[0-9]{6}$"
          description: 6-digit OTP code
          example: "123456"
...
    AuthResponse:
      type: object
      description: Successful authentication response with tokens and user info.
      required:
        - access_token
        - refresh_token
        - user
      properties:
        access_token:
          type: string
          description: JWT access token (HS256, 15 min TTL)
          example: eyJhbGciOiJIUzI1NiIs...
        refresh_token:
          type: string
          description: Opaque refresh token (64 hex chars, 7-day TTL), recorded against the requesting device's `X-Device-Id` refresh family
          example: a1b2c3d4e5f6...
        user:
          $ref: "#/components/schemas/User"
...
    AcceptedResponse:
      type: object
      description: Generic accepted response (always returned by OTP request).
      required:
        - status
      properties:
        status:
          type: string
          enum: [accepted]
          description: Always "accepted"
      example:
        status: accepted

    # ---------- Error schemas ----------
    ErrorResponse:
      type: object
      description: |
        Uniform error envelope used by all endpoints.
        The `details` field may contain additional context (e.g. validation errors).
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: invalid_body
            message:
              type: string
              description: Human-readable error description
              example: Invalid request body
            details:
              type: object
              description: Optional additional error context
              example: {}

    # ---------- Catalog & entries schemas ----------

    # Shared embedded tag (a category as it appears on an entry).
...